LegalDOCAI/backend/requirements.txt
faiss-cpu
pytesseract
pdf2image
Pillow
fastapi
uvicorn
transformers
torch
pymongo
